Started with index.txt impementation for tests
